requirements.txt
argopy 
pandas 
xarray
psycopg2-binary
chromadb 
langchain
langchain-community
langchain-openai
langchain-anthropic
langchain-huggingface
huggingface-hub
fastapi
flask
flask-cors
uvicorn 
duckdb
streamlit 
plotly
folium
sentence-transformers
python-dotenv
pyyaml
pytest
streamlit-fomantic